add list transfer schedules API

+++ schedule/transfer_detail.go
@@ -2,8 +2,8 @@ package schedule
 
 // TransferDetail represents transfer detail for schedule object.
 type TransferDetail struct {
-	Recipient           string `json:"recipient"`
-	Amount              *int   `json:"amount"`
-	PercentageOfBalance *int   `json:"percentage_of_balance"`
-	Currency            string `json:"currency"`
+	Recipient           string   `json:"recipient"`
+	Amount              *int     `json:"amount"`
+	PercentageOfBalance *float64 `json:"percentage_of_balance"`
+	Currency            string   `json:"currency"`
 }
